resin-init-flasher: Put board in suspend at the end of flashing
The watchdog timer will reboot the board after 128 seconds even after
a halt and this is problematic when using this board as a DUT in an
autokit setup, so let's instead put the board in suspend mode which
will also disable the watchdog. Going in suspend mode makes the ethernet
carrier disabled as expected so the autokit worker will now correctly
detect that the DUT has been successfully provisioned.

Changelog-entry: Get the board to enter suspend mode at the end of flashing
Signed-off-by: Florin Sarbu <florin@balena.io>